requirements.txt
python-telegram-bot==13.15
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
schedule==1.2.0
Pillow==10.3.0